fix: Model User
