+++ app/auth/routers.py
@@ -1,12 +1,11 @@
 from fastapi import APIRouter
 
-from app.auth.schemas import UserRead, UserCreate
-from app.auth.manager import auth_backend, fastapi_users
+from app.schemas.user import UserRead, UserCreate
+from app.core.user import auth_backend, fastapi_users
 
-api_version1 = APIRouter(prefix="/auth/jwt", tags=["auth"])
+auth_router = APIRouter(prefix="/auth/jwt", tags=["auth"])
 
-
-api_version1.include_router(fastapi_users.get_auth_router(auth_backend))
-api_version1.include_router(fastapi_users.get_reset_password_router())
-api_version1.include_router(fastapi_users.get_register_router(UserRead,
-                                                              UserCreate))
+auth_router.include_router(fastapi_users.get_auth_router(auth_backend))
+auth_router.include_router(fastapi_users.get_reset_password_router())
+auth_router.include_router(fastapi_users.get_register_router(UserRead,
+                                                             UserCreate))
