+++ app/core/config.py
@@ -1,13 +1,11 @@
-import os
-
 from pydantic_settings import BaseSettings
 
 
 class Settings(BaseSettings):
-    app_title: str = os.getenv('APP_TITLE')
-    description: str = os.getenv('DESCRIPTION')
-    secret: str = os.getenv('SECRET')
-    database_url: str = os.getenv('DATABASE_URL')
+    app_title: str = "AlfaBankHackathon"
+    description: str = "AlfaBankHackathon"
+    secret: str = "SECRET"
+    database_url: str = "sqlite+aiosqlite:///./fastapi.db"
 
     class Config:
         env_file = ".env"
