+++ app/crud/crud.py
@@ -2,7 +2,7 @@ from fastapi import Depends
 from fastapi_users.db import SQLAlchemyUserDatabase
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.models.base import User
+from app.models.user import User
 from app.core.db import get_async_session
